CORE: Fixed sponsored user deletion in impl layer
- We allow user connections like: "user -> sponsored -> service".
  This means, that "sponsored" user ID is stored in both DB columns
   (user_id, specific_user_id) as per relation (for his service, for his owner).
  When deleting sponsored user, we must delete all rows, where users id
  is in any of those two columns to remove both sides of relations.
- New version of deleteUser() can handle both normal and specific users.
- Removed deleteSpecificUser()

+++ perun-core/src/main/java/cz/metacentrum/perun/core/impl/UsersManagerImpl.java
@@ -354,21 +354,15 @@ public class UsersManagerImpl implements UsersManagerImplApi {
 		}
 	}
 
-	public void deleteUser(PerunSession sess, User user) throws InternalErrorException, UserAlreadyRemovedException {
+	public void deleteUser(PerunSession sess, User user) throws InternalErrorException, UserAlreadyRemovedException, SpecificUserAlreadyRemovedException {
 		try {
-			jdbc.update("delete from specific_user_users where user_id=?", user.getId());
+			// delete all relations like  user -> sponsor -> service
+			jdbc.update("delete from specific_user_users where specific_user_id=? or user_id=?", user.getId(), user.getId());
 			int numAffected = jdbc.update("delete from users where id=?", user.getId());
-			if(numAffected == 0) throw new UserAlreadyRemovedException("User: " + user);
-		} catch (RuntimeException err) {
-			throw new InternalErrorException(err);
-		}
-	}
-
-	public void deleteSpecificUser(PerunSession sess, User specificUser) throws InternalErrorException, SpecificUserAlreadyRemovedException {
-		try {
-			jdbc.update("delete from specific_user_users where specific_user_id=?", specificUser.getId());
-			int numAffected = jdbc.update("delete from users where id=?", specificUser.getId());
-			if(numAffected == 0) throw new SpecificUserAlreadyRemovedException("ServiceUser: " + specificUser);
+			if(numAffected == 0) {
+				if (user.isSpecificUser()) throw new SpecificUserAlreadyRemovedException("SpecificUser: " + user);
+				throw new UserAlreadyRemovedException("User: " + user);
+			}
 		} catch (RuntimeException err) {
 			throw new InternalErrorException(err);
 		}
